requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
Flask==2.2.5
Flask-JSON==0.4.0
flask-mongoengine==1.0.0
Flask-RESTful==0.3.10
Flask-WTF==1.2.2
frozenlist==1.8.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Markdown==3.9
MarkupSafe==3.0.3
mongoengine==0.29.1
multidict==6.7.0
ollama==0.6.0
packaging==25.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pymongo==3.12.0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
style==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
update==0.0.1
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
WTForms==3.2.1
yarl==1.22.0
